requirements.txt
flask
flask-cors
requests
python-dotenv
google-generativeai
mcp
openai